Comment out HRD with Stilts section
This section has been more or less superceded by the M4 with Stilts
section.

It would be nice to do something less textually disruptive
than actually commenting it out, but it's not easy to restructure
the document using macros and conditional inclusion because of the
verbatim environments which don't like being in macro arguments.
Could restructure it with external files and \input commands,
but for now keep it in one source file.