Figures for algebraic selection of comoving subset
I thought about drawing an ellipse when selecting the M4 comoving
sources in proper motion space.  The advantage is that the algebraic
expression (something like:
   hypot((pmra + 12.5)/4.4, (pmdec + 18.9)/3.2) < 1"
) can be reused in the STILTS part.  But it's a bit more complicated
to use/document, and it's nearly the first thing in the tutorial,
so for now leave it as it is.  But the figures are here if I change
my mind.